chore: 로그인 페이지 삭제, 사인인 페이지로 로직 이동.
