--- src/app/login/page.tsx
@@ -1,93 +1,0 @@
-'use client';
-
-import { FormProvider, useForm } from 'react-hook-form';
-import type { SubmitHandler } from 'react-hook-form';
-import { signIn } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
-
-interface LoginForm {
-  email: string;
-  password: string;
-}
-
-const LoginPage = () => {
-  const router = useRouter();
-  const methods = useForm<LoginForm>({
-    defaultValues: {
-      email: '',
-      password: '',
-    },
-  });
-
-  const onSubmit: SubmitHandler<LoginForm> = async (data) => {
-    try {
-      const result = await signIn('credentials', {
-        email: data.email,
-        password: data.password,
-        redirect: false,
-      });
-
-      if (result?.ok) {
-        router.push('/mypage');
-      } else {
-        console.error('로그인 실패:', result?.error);
-      }
-    } catch (error) {
-      console.error('로그인 에러:', error);
-    }
-  };
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-50 px-4 py-12 sm:px-6 lg:px-8">
-      <div className="w-full max-w-md space-y-8">
-        <div>
-          <h2 className="mt-6 text-center text-3xl font-bold tracking-tight text-gray-900">
-            로그인
-          </h2>
-        </div>
-        <FormProvider {...methods}>
-          <form
-            onSubmit={methods.handleSubmit(onSubmit)}
-            className="mt-8 space-y-6"
-          >
-            <div className="space-y-4 rounded-md shadow-sm">
-              <div>
-                <label htmlFor="email" className="sr-only">
-                  이메일
-                </label>
-                <input
-                  type="email"
-                  {...methods.register('email')}
-                  className="relative block w-full rounded-md border-0 px-3 py-1.5 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:z-10 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                  placeholder="이메일"
-                />
-              </div>
-              <div>
-                <label htmlFor="password" className="sr-only">
-                  비밀번호
-                </label>
-                <input
-                  type="password"
-                  {...methods.register('password')}
-                  className="relative block w-full rounded-md border-0 px-3 py-1.5 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:z-10 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                  placeholder="비밀번호"
-                />
-              </div>
-            </div>
-
-            <div>
-              <button
-                type="submit"
-                className="group relative flex w-full justify-center rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-              >
-                로그인
-              </button>
-            </div>
-          </form>
-        </FormProvider>
-      </div>
-    </div>
-  );
-};
-
-export default LoginPage;
