--- src/app/signin/page.tsx
@@ -1,9 +1,0 @@
-import SignIn from "@/components/SignIn/SignIn";
-
-export default function Page() {
-  return (
-    <div className="py-[167px]">
-      <SignIn />;
-    </div>
-  );
-}
